Moving Pydantic models for ECS components into component module
In order to simplify the work of referencing back to what attributes are available for use in the resource it's helpful to have the model definition inlined in the resource definition module.
