+++ tests/ol_infrastructure/components/fargate_service.py
@@ -7,12 +7,14 @@ from pulumi_aws.ec2 import (
     SecurityGroupIngressArgs,
 )
 
-from ol_infrastructure.components.aws.fargate_service import OLFargateService
+from ol_infrastructure.components.aws.fargate_service import (
+    OLFargateService,
+    OLFargateServiceConfig,
+)
 from ol_infrastructure.lib.aws.ecs.container_definition_config import (
     OLContainerLogConfig,
     OLFargateContainerDefinitionConfig,
 )
-from ol_infrastructure.lib.aws.ecs.fargate_service_config import OLFargateServiceConfig
 from ol_infrastructure.lib.aws.ecs.task_definition_config import (
     OLFargateTaskDefinitionConfig,
 )
